requirements.txt
beautifulsoup4
Google-Search-API
requests
pronouncing
sqlite3
Pillow
